main: Move chooseMetricsProvider to rollout.go
Since this function is only used by another function in rollout.go, it doesn't
make sense to have in main.go
